requirements.txt
# Requirements for Fake News Detection System (adjust versions as needed)
# Note: Some packages (torch, transformers) may require extra steps for GPU support.

Django>=4.2
streamlit>=1.24
transformers>=4.30
sentence-transformers>=2.2.2
torch>=2.0.1
scikit-learn>=1.2.2
pandas>=2.0
numpy<2.0.0
requests>=2.28
beautifulsoup4>=4.11
lxml>=4.9
Pillow>=10.2.0,<10.3.0
pytesseract>=0.3.10
openai-whisper>=20230314
python-dotenv>=1.0.0
tqdm>=4.65
openai>=1.0.0
pypdf
beautifulsoup4

# Database and Authentication
psycopg2-binary>=2.9.5
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
djangorestframework>=3.14.0

# Optional / production
gunicorn>=20.1.0
uvicorn>=0.22.0
